feat: export user role
